Add @JvmOverloads annotations where applicable; switch UInts to regular Ints

+++ source/sdk/src/main/java/com/stytch/sdk/b2b/oauth/OAuthImpl.kt
@@ -41,7 +41,7 @@ internal class OAuthImpl(
                 .authenticate(
                     oauthToken = parameters.oauthToken,
                     locale = parameters.locale,
-                    sessionDurationMinutes = parameters.sessionDurationMinutes.toInt(),
+                    sessionDurationMinutes = parameters.sessionDurationMinutes,
                     pkceCodeVerifier = pkce,
                     intermediateSessionToken = sessionStorage.intermediateSessionToken,
                 ).apply {
